config/cradio: Improve Bluetooth profiles (#59)
Limit the number of profiles to 4, the number of my devices (3) + 1 to
allow for a split keyboard to work. Bind buttons on the system layer to
switch to number profiles instead of navigating with next/previous as
it's quite annoying that I do not know which profile it's running. I've
also added buttons to disconnect from devices since it makes my phone
not show the keyboard.